+++ newIDE/app/src/MainFrame/EditorContainers/HomePage/TeamSection/TeamSection.spec.js
@@ -1,5 +1,5 @@
 // @flow
-import { groupMembersByGroupId } from '.';
+import { groupMembersByGroupId } from './utils';
 
 const getDefaultUser = ({ id }: { id: string }) => ({
   id,
