+++ newIDE/app/src/MainFrame/EditorContainers/HomePage/TeamSection/index.js
@@ -14,7 +14,6 @@ import CircularProgress from '../../../../UI/CircularProgress';
 import useForceUpdate from '../../../../Utils/UseForceUpdate';
 import {
   type TeamGroup,
-  type TeamMembership,
   type User,
 } from '../../../../Utils/GDevelopServices/User';
 import { type CloudProjectWithUserAccessInfo } from '../../../../Utils/GDevelopServices/Project';
@@ -34,6 +33,7 @@ import { ColumnStackLayout } from '../../../../UI/Layout';
 import Paper from '../../../../UI/Paper';
 import { useResponsiveWindowWidth } from '../../../../UI/Reponsive/ResponsiveWindowMeasurer';
 import RaisedButton from '../../../../UI/RaisedButton';
+import { groupMembersByGroupId } from './utils';
 
 const PADDING = 16;
 
@@ -50,8 +50,6 @@ const sortMembersByNameOrEmail = (a: User, b: User) => {
 
 const DropTarget = makeDropTarget('team-groups');
 
-type GroupWithMembers = {| group: TeamGroup, members: User[] |};
-
 type Props = {|
   project: ?gdProject,
   onOpenRecentFile: (file: FileMetadataAndStorageProviderName) => void,
@@ -61,51 +59,6 @@ type Props = {|
 export type TeamSectionInterface = {|
   forceUpdate: () => void,
 |};
-
-export const groupMembersByGroupId = ({
-  groups,
-  members,
-  memberships,
-}: {|
-  groups: ?(TeamGroup[]),
-  members: ?(User[]),
-  memberships: ?(TeamMembership[]),
-|}): ?{ [groupId: string]: GroupWithMembers } => {
-  if (!(groups && members && memberships)) return null;
-  const membersByGroupId = {};
-  members.forEach(member => {
-    const membership = memberships.find(
-      membership => membership.userId === member.id
-    );
-    if (!membership) return;
-    const memberGroups = membership.groups;
-    if (!memberGroups) {
-      const itemWithoutGroup = membersByGroupId['NONE'];
-      membersByGroupId['NONE'] = {
-        members: [
-          ...((itemWithoutGroup && itemWithoutGroup.members) || []),
-          member,
-        ],
-        group: { id: 'none', name: 'none' },
-      };
-      return;
-    }
-    const group = groups.find(group => group.id === memberGroups[0]);
-    if (!group) return;
-    const item = membersByGroupId[group.id];
-    if (item) {
-      item.members = [...item.members, member];
-    } else {
-      membersByGroupId[group.id] = { group, members: [member] };
-    }
-  });
-  groups.forEach(group => {
-    if (!(group.id in membersByGroupId)) {
-      membersByGroupId[group.id] = { group, members: [] };
-    }
-  });
-  return membersByGroupId;
-};
 
 const TeamSection = React.forwardRef<Props, TeamSectionInterface>(
   ({ project, onOpenRecentFile, storageProviders }, ref) => {
